Change email notifications to be once per assignment
This commit aims to change the email notifications sent only once a device is assigned or un-assigned. PUT and DELETE methods are idempodent and making the same requests should produce same results. So for example, once a device is assigned when it is tried to get re-assigned, it should not be sending email notifications